apps/remote_service_connector/pyproject.toml
[project]
name = "remote-service-connector"
version = "0.1.0"
description = "Modal-deployed connector for managing remote services (Cloudflare tunnels, SuperTokens auth) via authenticated HTTP endpoints"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.27.0",
    "imbue-common",
    "loguru>=0.7",
    "modal>=1.3.1",
    "paramiko>=3.0",
    "psycopg2-binary>=2.9",
    "pydantic>=2.0",
    "supertokens-python>=0.27.0",
]
license = {text = "FCL-1.0-MIT"}
license-files = [
    "LICENSE",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imbue"]

[tool.uv.sources]
imbue-common = { workspace = true }

# Shared pytest settings (markers, filterwarnings, coverage report config)
# are centralized in:
#   libs/imbue_common/imbue/imbue_common/conftest_hooks.py
# The settings below remain here because they are read from pyproject.toml
# before any hooks run: addopts, coverage.run, inline-snapshot, timeout_func_only.
[tool.pytest.ini_options]
timeout_func_only = true
junit_family = "xunit1"
testpaths = ["."]
addopts = [
  "--cov=imbue.remote_service_connector",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--durations=20",
  "--timeout=10",
  "--timeout-method=signal",
  "--slow-tests-to-file",
  "--coverage-to-file",
]

[tool.coverage.run]
parallel = true
concurrency = ["multiprocessing", "thread"]
omit = [
  "*_test.py",
  "test_*.py",
  "*/tests/*",
  "*/conftest.py",
  "*/testing.py",
]

[tool.inline-snapshot]
default-flags=["report"]
default-flags-tui=["report"]
default-flags-ide=["report"]
format-command="uv run ruff format --force-exclude --config pyproject.toml --stdin-filename {filename}"

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]

[tool.coverage.report]
fail_under = 45
